renamed repository variables, removed useless
